Make "config" API endpoint configuration agnostic
Even if we have been only providing Cognito configuration settings to
the front end, we could be providing any other type of configuration
settings. In fact, once we start using Keycloak with regular OpenID
Connect authentication flow, we will need to pass a different set of
values to the front end.

The endpoint code shouldn't care about what particular set of values
we are sending to the front end. It should just provide the map passed
in as a parameter, without assuming any kind of structure or set of
values.

Also adjust the front end to deal with the new config settings
structure.